Allow easier/possible Windows build
