+++ etpClientExample/MyOwnDirectedDiscoveryProtocolHandlers.cpp
@@ -35,7 +35,7 @@ void MyOwnDirectedDiscoveryProtocolHandlers::on_GetResourcesResponse(const Energ
 	std::cout << "*************************************************" << std::endl;
 
 	if (getObjectWhenDiscovered) {
-		Energistics::Etp::v12::Protocol::Store::GetObject getO;
+		Energistics::Etp::v12::Protocol::Store::GetObject_ getO;
 		getO.m_uri = graphResource.m_uri;
 		session->send(getO);
 	}
