+++ src/CMakeLists.txt
@@ -181,13 +181,19 @@ SET (ALL_SOURCES_AND_HEADERS
 )
 
 IF (WITH_ETP)
-	FILE (GLOB FESAPI_ETP_SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/etp/*.cpp ${CMAKE_CURRENT_SOURCE_DIR}/etp/ProtocolHandlers/*.cpp)
-	FILE (GLOB FESAPI_ETP_HEADERS ${CMAKE_CURRENT_SOURCE_DIR}/etp/*.h ${CMAKE_CURRENT_SOURCE_DIR}/etp/ProtocolHandlers/*.h)
+	add_compile_definitions (WITH_ETP)
+	
+	FILE (GLOB FESAPI_ETP_SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/etp/*.cpp)
+	FILE (GLOB FESAPI_ETP_PROTOCOL_SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/etp/ProtocolHandlers/*.cpp)
+	FILE (GLOB FESAPI_ETP_HEADERS ${CMAKE_CURRENT_SOURCE_DIR}/etp/*.h)
+	FILE (GLOB FESAPI_ETP_PROTOCOL_HEADERS ${CMAKE_CURRENT_SOURCE_DIR}/etp/ProtocolHandlers/*.h)
 
 	SET (ALL_SOURCES_AND_HEADERS
 		${ALL_SOURCES_AND_HEADERS}
 		${FESAPI_ETP_SOURCES}
-		${FESAPI_ETP_HEADERS}		
+		${FESAPI_ETP_PROTOCOL_SOURCES}
+		${FESAPI_ETP_HEADERS}	
+		${FESAPI_ETP_PROTOCOL_HEADERS}	
 	)
 
 	find_package (Threads)
@@ -207,10 +213,11 @@ IF (WITH_ETP)
 	IF (NOT IS_DIRECTORY ${BOOST_INCLUDE_DIR})
 		MESSAGE(WARNING "The BOOST include dir (BOOST_INCLUDE_DIR variable) does not look to be a valid directory. Please modify it.")
 	ENDIF ()
-	SET (BOOST_LIBRARY_RELEASE BOOST_LIBRARY_RELEASE-NOTFOUND CACHE FILEPATH "Path to the file which contains the BOOST library release")
-	IF (NOT EXISTS ${BOOST_LIBRARY_RELEASE})
-		MESSAGE(WARNING "The BOOST library (BOOST_LIBRARY_RELEASE variable) does not look to be a valid file. Please modify it.")
+	SET (BOOST_SYSTEM_LIBRARY BOOST_SYSTEM_LIBRARY-NOTFOUND CACHE FILEPATH "Path to the file which contains the Boost system static library release")
+	IF (NOT EXISTS ${BOOST_SYSTEM_LIBRARY})
+		MESSAGE(WARNING "The BOOST library (BOOST_SYSTEM_LIBRARY variable) does not look to be a valid file. Please modify it.")
 	ENDIF ()
+	add_compile_definitions (BOOST_ALL_NO_LIB)
 ENDIF (WITH_ETP)
 
 IF (WITH_JAVA_WRAPPING)
@@ -307,6 +314,7 @@ IF (WIN32)
 
 	IF (WITH_ETP)
 		SOURCE_GROUP ("${ETP_PREFIX}" FILES ${FESAPI_ETP_SOURCES} ${FESAPI_ETP_HEADERS})
+		SOURCE_GROUP ("${ETP_PREFIX}\\ProtocolHandlers" FILES ${FESAPI_ETP_PROTOCOL_SOURCES} ${FESAPI_ETP_PROTOCOL_HEADERS})
 	ENDIF (WITH_ETP)
 ENDIF (WIN32)
 
@@ -335,12 +343,12 @@ ENDIF (UNIX)
 # ============================================================================
 
 IF (WIN32)
-	ADD_DEFINITIONS (/bigobj)
+	add_compile_options (/bigobj)
 	IF (BUILD_SHARED_LIBS)
-		ADD_DEFINITIONS (-DFESAPI_DLL)
+		add_compile_definitions (FESAPI_DLL)
 	ENDIF (BUILD_SHARED_LIBS)
 	IF (HDF5_BUILT_AS_DYNAMIC_LIB)
-		ADD_DEFINITIONS (-DH5_BUILT_AS_DYNAMIC_LIB)
+		add_compile_definitions (H5_BUILT_AS_DYNAMIC_LIB)
 	ENDIF (HDF5_BUILT_AS_DYNAMIC_LIB)
 ENDIF (WIN32)
 
@@ -387,7 +395,7 @@ ELSEIF (UNIX)
 ENDIF (WIN32)
 
 IF (WITH_ETP)
-	TARGET_LINK_LIBRARIES (${CPP_LIBRARY_NAME} ${BOOST_LIBRARY_RELEASE} ${CMAKE_THREAD_LIBS_INIT} ${AVRO_LIBRARY_RELEASE})
+	TARGET_LINK_LIBRARIES (${CPP_LIBRARY_NAME} ${BOOST_SYSTEM_LIBRARY} ${CMAKE_THREAD_LIBS_INIT} ${AVRO_LIBRARY_RELEASE})
 ENDIF (WITH_ETP)
 
 # ============================================================================
