+++ src/etp/README.md
@@ -11,7 +11,7 @@ We advise you to install these third party libraries respectively into
  - give real path and files to the following cmake variables:
 	- BOOST
 		- BOOST_INCLUDE_DIR : where the "boost" directory containing all BOOST headers is located
-		- BOOST_LIBRARY_RELEASE : the BOOST system library you want to link to
+		- BOOST_SYSTEM_LIBRARY : the BOOST system library you want to link to
 	- MINIZIP
 		- AVRO_INCLUDE_DIR : where the "avro" directory containing all AVRO headers is located
 		- AVRO_LIBRARY_RELEASE : the AVRO system library you want to link to
