+++ src/tools/GuidTools.h
@@ -18,9 +18,34 @@ under the License.
 -----------------------------------------------------------------------*/
 #pragma once
 
+#if defined(_WIN32)
+#include <objbase.h>
+#endif
+
 #include <string>
 
 namespace GuidTools
 {
-	std::string generateUidAsString();
+#if defined(_WIN32)
+	inline
+#endif
+	std::string generateUidAsString()
+#if defined(_WIN32)
+	{
+		GUID sessionGUID = GUID_NULL;
+		HRESULT hr = CoCreateGuid(&sessionGUID);
+		wchar_t uuidWStr[39];
+		StringFromGUID2(sessionGUID, uuidWStr, 39);
+		uuidWStr[37] = '\0'; // Delete the closing bracket
+		char uuidStr[37];
+		wcstombs(uuidStr, uuidWStr + 1, 39); // +1 in order not to take into account the opening bracket
+		for (size_t i = 0; i < 37; ++i) {
+			uuidStr[i] = tolower(uuidStr[i]);
+		}
+
+		return uuidStr;
+	}
+#else
+	;
+#endif
 }
